Create XML reading function to populate a vector with InfoRefs
Created a function that takes a list of info refs and pushes the dereferenced info values into a vector, having cleared the vector first.

There are many examples of multi-level XML structures that can be simplified through the use of this structure.

<UnitAIs>
	<UnitAI>
		<UnitAIType>UNITAI_DEFENSE_AIR</UnitAIType>
		<bUnitAI>1</bUnitAI>
	</UnitAI>
	<UnitAI>
		<UnitAIType>UNITAI_CARRIER_AIR</UnitAIType>
		<bUnitAI>1</bUnitAI>
	</UnitAI>
</UnitAIs>

becomes

<UnitAIs>
	<UnitAIType>UNITAI_DEFENSE_AIR</UnitAIType>
	<UnitAIType>UNITAI_CARRIER_AIR</UnitAIType>
</UnitAIs>

Update XML reading code to use new function

Resolves: AB4-3