Correct conditional operators
The bitwise & operator was used instead of the logical && operator when checking if an event trigger was both a state religion and allowed a player to choose the religion. This meant that the condition would never fire.

The bitwise | operator was used instead of the logical || operator when working out the value of switching to a civic. This led to an incorrect value being assigned.

Resolves: AB4-2