Rehome units
If a unit ends its turn in a player city that is not its home city there is a chance that it will change its home city. This is potentially important when functionality is added to limit unit movement based on range from it's home city.

The rehoming chance is set as a global define.

AB4-4