Add plot improvement highlights from BUG/BULL
Added highlights to plots based on whether the plot is being worked by a city or not and whether the plot can be improved. All options, including highlight colours, configurable via in-game options screen.

Highlights are viewed on clicking the city name in the bar and not in the city screen itself.

AB4-12