Allow buildings to have civic prereqs
Buildings can be defined to only be constructed or operate when the player is running a specific civic or civics. The Civic(s) can be defined as OR or AND and if the player changes civics so that the prereqs are no longer met the building effects will be disabled. The building will still exist in the city and if the civic prereqs are met again at a later point the building will start functioning again.

AB4-16
AB4-2