Visually group prereqs in PlatyPedia
The PlatyPedia prereqs section of Units & Buildings object pages lists all the prereqs in a single row of icons. These are in categories such as tech, bonus or civic prereqs with the AND prereqs being listed first followed by the ORs. If there are more than one category then the display is similar to:

T1 T2 and (T3 or T4)

This works OK when there aren't many, but as the number and type of prereqs grows it can become messy, e.g.

T1 B1 and (B2 or B3) C1 C2 and (C3 or C4)

The use of the 'and' can get confusing as all the values are icons so it turns into

I1 I2 and (I3 or I4) I5 I6 and (I7 or I8)

A clearer mechanism is for the categories to be delimited by square brackets giving:

[T1] [B1 and (B2 or B3)] [C1 C2 and (C3 or C4)]

AB4-19