Refactor player initilization
The inclusion of the initInGame function by the Change Player mod introduced significant code duplication with the original player init funtion. These two functions were refactored to remove this. The original method signature remains as it is exposed to the exe, but points to a new combined function.

AB4-25
AB4-2