Allow techs to block improvement upgrades
A tech can be set that acts as a gatekeeper for the upgrade of an improvement. Presently improvements such as cottage/hamlet/village/town are upgraded based on the number of turns that the improvement is being worked.

The improvement must still meet the normal upgrade requirements. This should act to prevent upgrades too early in the tech tree.

AB4-36