Allow promotion groups
Allow promotions to be grouped so that once a unit has one promotion from a group it can swap between any promotion in the group for no XP expenditure once during each turn.

The promotion can define that it can only be applied when the unit is in a city.

AB4-43