Fix CTD in star events
The Gold modification star event was causing a CTD when attempting to report the effect for a civ with no capital city. Given that modifying gold is a civ wide change it made no sense to try and use the capital so this was updated to use the generic civ wide text.

AB4-74