Add world view concept
World views are the collective perception of what is acceptable for a player and are opened up by techs. These are large concepts such things as slavery that are decisions that are overarching in concept and not tied to things like civics which define how to achieve a goal rather than define the goal itself.

This change includes the framework, but not a full implementation of a specific world view. A partial implementation of the slavery world view is included to flesh out the framework, but the AI and specifics of what slavery enables are not included.

When a tech that newly enables a world view is researched human players are presented with a popup dialog as to whether they wish to start running the view or not. The AI has a routine to decide for them.

In order to change a world view after the tech is first researched requires a GP and there is a time restriction between changes that is based on the game speed and a global variable. The AI should have a routine to decide whether the world view is a positive or negative for them and toggle the view appropriately. A human will receive a popup with all know views and they can toggle one of these. This does not consume the GP.

The concept and some of the code came from the C2C mod

AB4-78