Show SubCombat type icons
The icons for sub combat types are now shown on the plot info screen when a unit is selected and in the pedia entry for the unit.
In the PlatyPedia units were only shown in their primary unit combat type and not for any sub types, they are now shown under all combat types that they belong to.

AB4-83