Prevent CTD in WorldBuilder CityEdit screen
Opening the city edit (city data 1) screen in the world builder by selecting a city would cause the game a to CTD. Attaching a debug dll indicates that the exe is trying to delete a non existent object. This has been tracked down to the use of the city order queue, but the exact cause is still unknown.

As a workaround the code to read the city queue on screen opening has been commented out whilst the root cause is determined.

AB4-90