Rework Natural Wonders
There are several places where natural wonders may by replaced on the map during the map generation process. This not only has the effect of removing the feature, but does not remove the additional plot yields leaving some plots in an unwanted state. The solution is to ensure that once placed the NW remains.

The Natural Wonders are placed on the map before normal features to allow multi-plot wonders to find a viable position more easily. The normal features being placed afterwards did not check whether there was a NW on a plot so could overwrite it. They also did not remove the extra plot yields that come with the NW. This has been fixed so that a plot is not considered valid for a feature if there is a pre-existing feature there.

The routine that removed bad features from around the civ starting plots treated natural wonders as bad as they have no inherent yield and removed them accordingly. This did not remove the extra yield produced when the feature was initially placed leaving a plot with an incorrectly high yield.

Some natural wonders spread over multiple plots and players would receive the movie and/or notification each time they revealed one of the plots. This change ensures that a player only receives a single notification for a natural wonder. Added the movie definitions so that the movies show

Removed NW buildings from the tech tree and pedia

AB4-65