Refactored spy unit variables to new class
There were a significant number of CvUnit variables that were spy specific. Given the very large number of units for which these variables were not applicable they were bloating the memory footprint for minimal gain. Refactored these variables to a new class, CvSpy, which is accessible via composition. An instance of the CvSpy class is only created for spy units.

AB4-101