Allow abandoning City/Raze Building
Whilst in the city screen Ctrl-A will bring up a dialog allowing you to either demolish buildings or abandon the city completely. If you abandon the city it is considered razed with all the associated implications.

On abandoning a city you receive the latest kind of settler you can build for each 10 population with each remaining 3 population producing a worker.

On demolishing a building you receive 20% (scaled by game speed) of the production costs back in gold, with obsolete building having a further 50% reduction.

The values are configurable in an ini file

Based on code by Bhruic, Nemo, OrionVetran & Dancing Hoskuld

AB4-111