Show graphical only buildings in the city screen
Buildings with the bGraphicalOnly tag are not normally shown in any of the UI elements (Pedia or City Screen). There are some buildings, such as those associated with natural wonders, that should not be shown in the pedia, but are useful to show on the city screen to indicate where some city changes are coming from.

This commit adds an additional tag that will override the bGraphicalOnly tag on the city screen only.

AB4-114