Add Wait for tech mission to Great People
Add a mission that allows great people to select a tech that they can discover at a later time and sleep until then. Once they can discover the tech they wake and bulb it. Only techs that are appropriate for the GP type are available based on their flavour values

AB4-117