Added pillage automation
Added a new automation for units to pillage improvements

There were two AI pillage routines one that checked every map plot and one that checked a range of tiles around the unit. Given that checking every plot is inefficient on large maps I have removed that and changed calls to it to use a large range instead. This will hopefully improve performance as well as getting rid of duplicated code.

AB4-118