Add function to indicate if plot is subject to a fort attack
There was currently no function to determine whether a unit on a plot would be a potential target from a fort attack. This had implications for the AI_heal and other routines that would potentially be stuck in an endless heal/take damage cycle.

The function to determine if a plot can fort attack has also been updated to only return true if there is a valid attacking unit present.

I have not included the new method anywhere at this point.

AB4-119
AB4-82