Add City Defender automation
Add a unit automation to protect the city from nearby attackers. The main priority is to stay close to the city and if it is the best defender then stay in the city itself. If there are other defenders then it will look within the city bounds for enemies to attack.

AB4-122