Sort promotions by UnitCombat in PlatyPedia
The PlatyPedia has an option to display promotions sorted by UnitCombats. When selecting this view nothing is displayed due to a change to the way that UnitCombats are handled in the XML file. To fix this the call to Promotion.getUnitCombat should be changed to Promition.isOrCombatType in PlatyPediaMain.py

AB4-135