Improvements can spawn units
Improvements can spawn units periodically. On creation of the improvement a non-mobile unit of an appropriate type with additional power is created on the plot as a guardian. If the guardian is killed the next unit spawned will become another guardian.

Improvements can be created that act as lairs for animals and barbarians, these are created at the appropriate time in the game when these spawn.

Added an improvement that spawns animals periodically based on the underlying terrain

AB4-139