Add Slavery concept
Any unit type may perform a one-way conversion to a slaver. Slavers have their own AI that hunts out vulnerable units and on either capture or kill enslaves them. Once a unit has been enslaved it may be taken to a city with a slave market building and sold to produce a slave.

Slaves may either be put to work in the city producing a minor food and productivity boost or used for builds outside the city. If used for a build it will kill the slave. If a slave is settled in a city there is a chance each round that the slaves will rebel based on the number of slaves vs normal population. Military units in a city help prevent slave unrest.

Slavers are invisible and a promotion is available that is able to see them.

The AI remembers when it has been the target of slavers on an area by area basis. If it has been targetted it will send escorts for its workers. The AI will product slavers based on the number of other civs in the area and the number of existing slavers it has there. The AI also knows how to send slavers out to hunt for slaves and how to select a city to sell them at along with how the slaves maybe best used.

The slavery concept is enabled/disabled via a concept known as a world view that is available at Bronze Working. If the world view is repealed then any slavery based units and buildings will stop working until the view is enabled again. The more slaves a civ has the more chance there is for a slave revolt and this is done on a city by city basis. If a civ researches the Liberation tech then the chance of slave revolts around the world increases. The AI has routines that help decide when to repeal the slavery world view based on the chance of revolt in its cities.

AB4-141