Provide yield to buildings with bonus
Allow buildings to define a yield bonus based on whether the city has bonus resource available.

The bonus can be optionally connected or also in it’s city workable radius. The bonus should be able to be predicated on whether the bonus merely exists or whether the bonus must be being actively provided to the city via an improvement and trade network.

AB4-150