Add a tag to help sort civics in java tooling
The order of the entries in the civics xml file determines how they are displayed on the civic screen in game. Without any guidance the java tooling has no way of knowing the correct order to when creating the file as a new tag has been created for this purpose. The tag is not read in the dll and is for sole use of the tooling

AB4-147