Reorder entries in BuldingClassInfos file
The order of entries in this file determines how constructed buildings are display in the city screen and also the order they are displayed when selecting a new building to construct in the same screen. Without the reordering there was no structure making it harder to identify good choices when the number of buildings increases.

This commit includes the updated XML and schema file

AB4-152
AB4-147