Have city construct food yield production buildings when not growing
The city ai was not fully taking into account the direct food production of some buildings when it was looking to grow, this was causing it to ignore this feature of buildings when selecting a production target. As a result city growth would be stagnated where it could be growing. This commit adds a value for this building effect when the city is looking for growth.

AB4-169