Show vicinity bonus and building prereqs on the PlatyPedia Building page
These prereqs were shown in the hover text for buildings but excluded from the building specific pedia page. This commit ensures they are shown on the building specific pedia page.

The python code has been updated to show building terrain prereqs in the icon "Required" section rather than the "Special Abilities" as these are more appropriate there. This decision may need to be revisited if it turns out that there are too many icons to show for some buildings, but the terrain prereqs are uncommon and buildings that have them should not have too many other icon displayed prereqs

AB4-170