Techs can modify the flat yield that a building provides to a city
The yield that a building provides to a city can be modified by researching techs as a representation of more efficient processing methods. This can be used as an alternative to obsoleting one building with another. It means that the players get a ‘free’ upgrade rather than requiring a new building to be constructed to gain the benefit so may not be appropriate in all cases as typically a player should have to work to gain benefits in a city.

AB4-171