Allow trait to modify the players gold by a percentage each turn
At the end of each turn a player with this trait will have their gold modified by the given percentage. It is expected that this figure is kept small otherwise this could be very overpowering, but two or three percent should not be too strong yet still add up over the course of a game. It also requires the player to have excess gold before its effect is felt.

AB4-177