Allow trait to remove religious and civic anarchy independently
OOB traits modify both religious and civic anarchy as sinlge entity but this is not really the case and the Spiritual trait should not be preventing civic anarchy. This commit adds a couple of new tags that allow these to be prevented independently.

The existing tag to modify both traits has been removed and the Spritual trait updated to use both new tags so as not to change its functionality at this point.

AB4-183