Features may require to be adjacent to water
Add tags to features to enforce that they are adjacent to a water area of a given minimum and maximum and maximum size. This is to ensure that the Dead Sea natural wonder building which improves water tile yields has a value but sill hopefully not apply to sea or ocean tiles.

Add this tag to the Dead Sea NW with a min water area size of 1 and maximum of 4.

AB4-65
AB4-187