Sorting Civics by CivicOptionType in pedia should not use an alpha-sort
In the Pedia it is possible to show the civics using multiple sorting styles, one of these is to sort them by the CivicOptionType. This sort currently has the civics within each type alpha sorted but it is more useful to show them in the order they appear in the XML file as this allows the order that they are available to be used instead, assuming that the XML file is created that way.

AB4-191