Buildings can apply promotions to units ending their turn in city
A building that applies a promotion to new units can have an additional configuration that allows it to apply the promotion to units that end their turn in the city as well. This should be used sparingly as the free promotion is intended to be incorporated into training period.

A building should also be able to be configured to apply a promotion on end turn for any building that does not have its own end of turn configuration. This means that any building that provides a free promotion on creation will also apply it on unit end of turn irrespective of whether it is normally configured to do so or not.

AB4-194