Allow buildings to add culture to cities if owned units die in them
Buildings should have the ability to have the death of one of their owners units in their city, or all player cities, generate culture for the city. This should also negate any Influence Driven War culture updates. The culture generated should be equal to the level of the unit that died.

AB4-195