Add concept of Unit Morale
Each unit has an individual morale value based on the Unit Type, this can be modified through the use of promotions and also the damage that the unit has taken. The morale change via damage is based on the square of the damage so that at low damage there is minimal change to morale, but once significant damage has been taken the unit morale may be significantly impacted. Once a configurable damage has been taken the unit is subject to morale checks and if it fails then it is considered to have been routed and the battle ends. If the unit is mechanised then it may leave its equipment behind for the victor.

If a unit has been routed mid-battle then it is still considered to be dead, but the victor gains a benefit of not having any more rounds where they may lose damage so could come out of the battle with more health than if the unit had fought to the death.

Morale checks are done against a base value of 100; if a units morale is higher than this when checked then it will not break. Enemy units may gain promotions that modify (percentage change) the morale of units they are attacking so it is possible for any unit to break.

The default morale of a unit type is 750 and with this value they are very unlikely to rout unless a battle round leaves them with very low health or they are facing an enemy with a modifier.

AB4-196