Add unit plunder concept
Attacking Units can plunder gold from an enemy civ on combat victory. The gold earned is removed from the enemy civs reserves and added to the attacking players gold. There is a cap on the total gold that may be plundered from each civ per turn of 10%. Each player can gain up to this amount in plunder from a single civ, but the target civ will only ever lose the amount once. This is a compromise that has the effect of making money. without it if each attacking civ could take 10% then the target civs gold could be wiped out in a single turn and if the amount to be plundered were only available once then the order that the civs are processed would determine which civ got the gold which could make the ability redundant fairly quickly.

Each attacking civ can only take up to the cap of gold from a single target civ and any plundering units that win battles afterwards gain a single gold, but it is not taken from the target civs reserves.

The amount that can be plundered from a civ is recalculated each turn along with the amount that each attacking Civ has taken.

AB4-197