Add Weapon and Ammo material types
The material that a units weapons are made of impacts the unit strength. Only some, typically early, unit combat types are impacted by weapon material whilst others are unaffected. The materials can update melee weapons, bows, siege weapons and also projectiles such as arrows and bolts.

The UnitInfo can define the maximum weapon/ammo tier that is available to that unit.

The materials should be defined in their own XML file and contain details about which unit combats can be affected, any bonus that is required and the additional strength it gives to units wielding it.

The main interface should show unit weapons in the unit info display

If a unit ends its turn in a city that has access to better quality weapons or Ammo the unit will automatically upgrade to use the best available.

The UI displays an icon for any unit that has an upgraded weapon and the possible upgrades are shown in the pedia for units.

AB4-204