Weapon types don't always upgrade
Some weapon type upgrades are inherent in the construction of the unit and cannot be upgraded. It should be possible to define these weapon types to allow for the upgrade on training the unit, but not afterwards. An example would be siege units or the hulls of wooden ships.

AB4-204