Limit early unit exploration
In the ancient era units should not be free to explore the complete map so a range limit has been set on the early units in this era. It is not intended that this limit should be in place too long so later eras remove all restrictions. Units that have a range of UNITRANGE_HOME are still restricted to their home city plots.

AB4-212