Limit strong animal improvement spawning
In the early game having an animal lair that spawns strong animals such as Tigers can give a player a distinct disadvantage as these cannot be killed by starting units and as they keep spawning without having their numbers reduced it can become near impossible for units to scout the area.

This commit addresses this by assuming that stronger animals require a larger territory to operate in and therefore would require more free tiles before they will spawn an additional unit. The stronger the animal the more free tiles are required for the lair to spawn.

AB4-139
AB4-213