Update Future Era to adjust to Transhuman
When the updated tech tree was included as part of AB4-207 in commit a05cf648 the Transhuman era was added, but the Future era was not updated to accommodate this with regard to the AdvancedStartCost value. The java tooling uses this tag value to sort the entries and as they were both the same only one of the eras was imported from the spreadsheet.

This commit updates the values and includes the latest toolkit export spreadsheet

AB4-207