Add Law Enforcement units
Law enforcement units are a new combat class for city defence. The units are all bound to the cities BFC and are only defensive. They do not have military support costs and have a bonus to city defence so are well suited to their primary function.

The following units have been added
- Slinger
- Fletcher
- Minuteman
- Police

AB4-215