Add ability to show Great General progress bar
Currently only the Great Person progress is displayed on the main screen. Added functionality to also show the GG progress. These bars are dynamically handled so that none, 1 or more are able to be displayed and the size of each bar alter based on the number shown.

The options dialog has values that can be set to scale each progress bar relative to the others. A progress bar is only be displayed if there is a non-zero value for the progress.

AB4-217