Update city growth factors
Cities should not be able to grow rapidly in the early eras. In order to slow their growth the food that each pop requires has been increased. There are additional food generating building available with this mod (or there will be) to balance this out.

With the addition of multiple buildings that directly provide food the food that a city needs to grow needs to be extended to prevent rapid growth in the early game and snowballing later on.

AB4-219