Rebalance Natural Wonders
Having a natural wonder close to the capital gives too large a culture boost, negating the point of buildings such as the Monument. This commit reduces the culture gained from 4, and in some cases higher, down to 1. The other impacts have also been rebalanced somewhat so that all wonders have a roughly equal impact if played into.

AB4-65