Add new specialists
New city specialists added to allow more control over a city focus. There are no current specialists that allow for management of the Food yield or happiness so options for these should be available.

Three new specialists should be added:

Farmer - increases Food yield by 3 at the expense of a Production allowing for early growth in cities where there is limited natural food production. This is necessary as some starting locations make it difficult for cities to grow given that citizen food consumption has been increased to 3 for the game.

Noble - similar to Merchants in that they increase the gold of a city, but they are less effective but also give a small boost to culture.

Jester - adds happiness to city along with some culture at the expense of food production. The Jester will eventually provide GPP towards a Great Jester GP, but this unit has not been implemented yet.

AB4-235