Correct game speeds for no stone age
The game speed file was originally created for when there were stone age eras added at the start with an expansion of the tech tree prior to the ancient era. This meant that the game year as indicated in the game main screen was tens of thousands of years out very quickly. This commit corrects the game year as reported.

AB4-145