Allow settlers to ignore resident animals
If their is a spawning lair near to a proposed city site the AI would not send their settlers there due to the danger and the settler could end up toing and froing in an attempt to get to the spot and also avoid danger. In order to try and resolve this resident animal are set as non-hostile, allowing the settler to pass.

The calculation of the number of free tiles, representing animals hunting/foraging range, required has been increased by 1 as low combat animals were being spammed to the point that there could be little room if they were blocked by a city BFC.

AB4-139
AB4-213