Add Brightsmiths hut
The Brightsmiths hut is an early commerce building that requires Gold/Silver/Gems in the city radius and will produce a small amount of commerce for each. The resource does not need to be connected as this building is available before the trade network for these is available but should not be overpowered as not many cities will meet the resource requirements.

AB4-238