Add Ancient Mysteries national wonder
On construction of this wonder the player is able to select an additional trait for their leader. The functionality is implemented in Python so the AIWeight building tag been used to allow the AI to evaluate this. This tag has been given a very high value as the addition of a new trait is something that should be prioritised in nearly all cases given it’s overarching benefit.

AB4-283