Allow settlers to ignore resident animals
The original fix for this tried setting resident animals as non-hostile, but this did not fix the issue. Now animals resident in a lair are set as defensive only which means that the the code to check for a hostile plot now returns false, allowing settlers to pass.

AB4-139
AB4-213