Update civ starting techs based on new tech tree
Some of the OOB civ starting techs were in the second tier and these meant that these civs would never be able to gain the benefits from the prereq techs, but were able to immediately gain the benefits of the ter 2 techs. This commit updates all civs to only be given tier 1 techs for free.

AB4-207