Add Trade Caravan unit
The trade caravan allows high production cities to transfer some of their hammers to other cities to allow those cities on more marginal plots to still grow and construct buildings. It uses the same AI routines as the Engineer & Merchant GPs which will allow it to intelligently select cities to help improve.

AB4-311