Ensure Java tooling XML output keeps UBs with parent
The standard sorting of infos uses an alpha sort which works most of the time, but there are some scenarios where a UB would be out of order if the parent starts with the same text as another info. An example would be for the following buildings:

Walls
Walls: Celtic Dun
Walls Of Troy

Using standard alphasort the Walls of Troy would be output before the Celtic Dun due to space having a lower ASCII value than the colon.

This commit contains a custom comparator that treats a colon as coming before a space when this is the first character that differs between two strings

AB4-319