Add Devotee Retinue promotion
This promotion is applicable to settlers which spreads any state religion to new cities on founding. The promotion can only be given through buildings or other system configurations, it is not possible for the player to directly select it.

This commit does not apply the promotion and it needs to be added to buildings or other means.

AB4-329