Add missionary activation effects
When a missionary spreads a religion the religion icon should be displayed from the unit. This was missing from the missionaries of the new religions and has been added in this commit.

Details of how to link the EffectInfo with the unit animation are described here: https://forums.civfanatics.com/threads/quick-modding-questions-thread.351519/page-320#post-16324949

AB4-320
AB4-334