Remove circular building prereq dependencies
Some buildings were defined as replacements for another building that was also a prereq. This led to an undefined state which ended up with the game preventing any building being constructed in a city. This commit also revisits some of the other building prereqs to determine if they were actually valid or just there because they sounded good.

AB4-366